fixed the compile error caused by gtk_combo_box_text_new()
GtkComboBoxText was implemented in GTK since 2.24, while in some
Linux Distro such as Ubuntu10.10 or older versions, on which the
GTK version is 2.22 or old, it would cause a compile error.

I reimplemented the combo box using gtk_combo_box_new_text()
instead of gtk_combo_box_text_new() to solve this problem.

Signed-off-by: levin li <levin108@gmail.com>